requirements.txt
pygame>=1.9.6
pyYAML>=5.3
pyqt5
numba>=0.50.0 
